updated to python 3, did some other code cleanup

+++ layers/reorg/reorg_layer.py
@@ -1,6 +1,7 @@
 import torch
 from torch.autograd import Function
-from _ext import reorg_layer
+from ._ext import reorg_layer
+import numpy as np
 
 
 class ReorgFunction(Function):
@@ -26,7 +27,7 @@ class ReorgFunction(Function):
         stride = self.stride
         bsize, c, h, w = grad_top.size()
 
-        out_w, out_h, out_c = w * stride, h * stride, c / (stride * stride)
+        out_w, out_h, out_c = w * stride, h * stride, int(c / (stride * stride))
         grad_bottom = torch.FloatTensor(bsize, out_c, out_h, out_w)
 
         # rev_stride = 1. / stride    # reverse
